update tests uodate readme.md
