--- requirements.txt
@@ -1,2 +1,0 @@
-kaggle==1.6.17
-requests
